Use uintptr_t for chipaddr instead of unsigned long.
Acked-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>

git-svn-id: svn://flashrom.org/flashrom/trunk@1698 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1

+++ flash.h
@@ -41,7 +41,8 @@
 #define ERROR_OOM	-100
 #define TIMEOUT_ERROR	-101
 
-typedef unsigned long chipaddr;
+/* TODO: check using code for correct usage of types */
+typedef uintptr_t chipaddr;
 
 int register_shutdown(int (*function) (void *data), void *data);
 void *programmer_map_flash_region(const char *descr, unsigned long phys_addr,
